Use scripts for controlling interactions with objects.
Add the Gemini script compiler and virtual machine.

Add the sample movement script. Hook it up to the space key.
The sample movement script exercises StartTrack, which starts a script for another object - so that several NPCs can move at the same time.

Include all of the code that deals with scene scripts conditionally with a preprocessor macro.