Implement type checking
Add type denoters for pointer, function, and named types like int.
The clauses of a If/Cond/Case must have compatible types.
Break, Next, and Return have a type that can be superseded.
This type is called Xfer (for "transfer of control")

TypeRefs have a type (of type Type), and a referent type,
which is what is defined by the type ref.